feat(proxy): default to v2 migration resolver, add --use_legacy_migration_resolver
The v2 migration resolver — which skips the diff-and-force recovery that
caused schema thrashing when two LiteLLM versions contend for the same DB
during rolling deploys — is now the proxy default.

- `--use_v2_migration_resolver` is kept as a deprecated no-op that warns
  (so existing scripts don't break on an unknown-option startup error).
- `--use_legacy_migration_resolver` opts back into the v1 resolver.

Testing:
- Migrate the resolver unit tests into the standard suite at
  tests/test_litellm/proxy/db/test_migration_resolver.py (they previously
  lived in litellm-proxy-extras/tests, which runs in no CI) and add
  legacy-resolver coverage alongside the existing v2 tests.
- Add proxy_cli unit tests for the new default, the deprecated no-op, and
  the legacy opt-out.
- Repoint the dedicated Postgres CircleCI job at the legacy resolver so v1
  keeps real-DB + proxy-boot coverage (v2 is the default everywhere else).
